Rename directories and chnage namepsane in Database/
