+++ app/Database/Connection.php
@@ -1,21 +1,22 @@
 <?php
 
-    namespace Connection;
+    namespace Database;
     
     // import global variable
     use mysqli;
 
     // declare static variable for connection with database
-    define("DB_HOST", "localhost");
+    define("DB_HOST", "127.0.0.1");
     define("DB_USER", "root");
     define("DB_PASSWORD", "");
     define("DB_NAME", "week_planer");
 
     class Connection {
         static public function connect() {
-            // 
-            $con = new mysqli("127.0.0.1", "root", "", "week_planer");
+            // create connection with database
+            $con = new mysqli(DB_HOST, DB_USER, DB_PASSWORD, DB_NAME);
 
+            // handle errors
             if($con->connect_error) {
                 echo "Error with connection with db";
             }
